move new test to existing file, allocate map for struct up front
